requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioresponses==0.7.6
aiosignal==1.3.2
attrs==24.3.0
bitstring==4.0.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.7.0
defusedxml==0.7.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
pillow==11.1.0
propcache==0.2.1
pybars3==0.9.7
pycparser==2.22
PyMeta3==0.5.1
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.31.0
responses==0.25.0
tinycss2==1.4.0
urllib3==2.3.0
webencodings==0.5.1
yarl==1.18.3
yattag==1.15.1
pytest==8.1.1
pytest-asyncio==0.23.5
